removed file used for testing, updated anonymization, moved logging position
